// support/enums/productNamesEnums.ts
export enum ProductNames {
    BlueAndGreenSneaker = "/blue-and-green-sneaker",
    BlueJeans = "/blue-jeans",
    BuildYourCheapComputer = "/build-your-cheap-own-computer",
    BuildYourOwnExpensiveComputer = "/build-your-own-expensive-computer-2",
    BlackAndWhiteDiamondHeart = "/black-white-diamond-heart",
    BookComputingAndInternet = "/computing-and-internet"
}